// src/app/customer/models/customer-real-owner.ts
export interface CustomerRealOwner {
  id?: number;
  personId?: number;
  curp: string;
  foreignerWithoutCurp: boolean;
  rfc: string;
  firstName: string;
  middleName: string;
  dateOfBirth: string;
  firstLastName: string;
  secondLastName: string;
  gender: string;
  maritalStatus?: string;
  nationality: string;
  countryOfBirth: string;
  stateOfBirth: string;
  countryTaxCodeAbroad?: string;
  typeIden: string;
  typePerson?: string;
  relationship: string;
  field: string;
  phone: string;
  mail: string;
  economicActivity: string;
  expirationDateField: string;
}

export interface CustomerRealOwnerData {
  generalData: CustomerRealOwner;
  ppe: CustomerRealOwnerPPE;
  adrres: CustomerAddress;
}

export interface CustomerAddress {
  id?: string | number,
  addressId?: number,
  addressAccountId?: number,
  addressRole?: string,
  addressType: string,
  other?: string,
  country: string,
  postalCode: string,
  federalEntity: string,
  city: string,
  municipality: string,
  neighborhood: string,
  street: string,
  externalNumber: string,
  internalNumber?: string,
  confirmCp?: string,
  timeLiveMexico?: string,
  reasonsOpeningContractMexico?: string,
  proofOfAddressType?: string,
  addressProofIssueDate?: string,
  expirationYear?: string,
  taxPostalCode?: string,
  geographicalArea?: string,
  deliveryCenter?: string,
  neighborhoodName?: string,
  addressConcatenation?: string,
  federalEntityID?: string,
  cityID?: string,
  municipalityID?: string,
  isSaved?: boolean,
  isView?: boolean,
  active?: boolean,
}
export interface CustomerRealOwnerPPE {
  idM?: number;
  id?: number;
  ppe: boolean;
  tppe: string;
  positionHeld: string;
  expirationDate: string;
  fppe: boolean;
  dataFamily: Array<DataRealOwnerClientFamilyPPESave>;
}

export interface DataRealOwnerClientFamilyPPE
  extends CustomerMinRealOwnerDataClient,
  CustomerPositionHeld {
  idM?: number;
  id?: string | number;
  idS?: string;
  accountRoleId?: number;
  active?: boolean;
  personId?: number;
  isView?: boolean;
  isSaved?: boolean;
}

interface DataRealOwnerClientFamilyPPESave
  extends CustomerMinRealOwnerDataClient,
  CustomerPositionHeld {
  idM?: number;
  id?: string | number;
  idS?: string;
  personId?: number;
  active?: boolean;
  accountRoleId?: number;
  isView?: boolean;
  isSaved?: boolean;
}